Move model file from mvc/library
